# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CalCard JSON API
  version: 0.1.0
  description: |
    OpenAPI documentation for CalCard's JSON calendar endpoints under `/api`.
    CalDAV and CardDAV endpoints under `/dav` are protocol endpoints and are not
    represented here.
servers:
  - url: /
security:
  - basicAuth: []
...
paths:
...
  /api/addressbooks:
    get:
      tags:
        - Address Books
      operationId: listAddressBooks
      summary: List address books
      responses:
        "200":
          description: Address books owned by the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AddressBook"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/addressbooks/{id}:
    parameters:
      - $ref: "#/components/parameters/AddressBookID"
    get:
      tags:
        - Address Books
      operationId: getAddressBook
      summary: Get an address book
      responses:
        "200":
          description: Address book metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressBook"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/addressbooks/{id}/contacts:
    parameters:
      - $ref: "#/components/parameters/AddressBookID"
    get:
      tags:
        - Contacts
      operationId: listContacts
      summary: List contacts in an address book
      description: |
        Returns contacts in the address book. Optional query parameters filter
        the results in the database (scoped to the address book, so no full table
        scan). Text filters are case-insensitive substring matches and are
        combined with AND. Results are ordered by display name.
      parameters:
        - name: name
          in: query
          required: false
          description: Case-insensitive substring match on the display name.
          schema:
            type: string
            maxLength: 256
        - name: email
          in: query
          required: false
          description: Case-insensitive substring match on the primary email.
          schema:
            type: string
            maxLength: 256
        - name: q
          in: query
          required: false
          description: Case-insensitive substring match across display name and email.
          schema:
            type: string
            maxLength: 256
        - name: limit
          in: query
          required: false
          description: Maximum number of contacts to return. Values above 1000 are capped.
          schema:
            type: integer
            minimum: 0
            maximum: 1000
        - name: offset
          in: query
          required: false
          description: Number of contacts to skip, for pagination.
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Contacts in the address book.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Contacts
      operationId: createContact
      summary: Create a contact
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IfNoneMatch"
      requestBody:
        $ref: "#/components/requestBodies/ContactWrite"
      responses:
        "200":
          description: Existing contact updated by the create request.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "201":
          description: Contact created.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/addressbooks/{id}/contacts/{uid}:
    parameters:
      - $ref: "#/components/parameters/AddressBookID"
      - $ref: "#/components/parameters/ContactUID"
    get:
      tags:
        - Contacts
      operationId: getContact
      summary: Get a contact
      responses:
        "200":
          description: Contact details and raw vCard data.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Contacts
      operationId: updateContact
      summary: Update a contact
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IfNoneMatch"
      requestBody:
        $ref: "#/components/requestBodies/ContactWrite"
      responses:
        "200":
          description: Contact updated.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Contacts
      operationId: deleteContact
      summary: Delete a contact
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "204":
          description: Contact deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: DAV Basic authentication using the user's email and an app password.
  parameters:
...
    AddressBookID:
      name: id
      in: path
      required: true
      description: Numeric address book identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
    ContactUID:
      name: uid
      in: path
      required: true
      description: URL-encoded vCard UID.
      schema:
        type: string
        minLength: 1
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: Only apply the write if the current event ETag matches.
      schema:
        type: string
      example: "\"abc123\""
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: Only apply the write if the current event ETag does not match.
      schema:
        type: string
      example: "*"
  headers:
    ETag:
      description: Current entity tag for the event.
      schema:
        type: string
      example: "\"abc123\""
  requestBodies:
...
    ContactWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContactWriteRequest"
        text/vcard:
          schema:
            $ref: "#/components/schemas/RawVCard"
        application/vcard:
          schema:
            $ref: "#/components/schemas/RawVCard"
  responses:
    BadRequest:
      description: Invalid path parameter, request body, calendar data, or structured event data.
      content:
        text/plain; charset=utf-8:
          schema:
            $ref: "#/components/schemas/ErrorText"
    Unauthorized:
      description: Missing or invalid DAV Basic authentication.
      headers:
        WWW-Authenticate:
          description: Basic authentication challenge.
          schema:
            type: string
          example: Basic realm="CalCard DAV"
      content:
        text/plain; charset=utf-8:
          schema:
            $ref: "#/components/schemas/ErrorText"
...
    NotFound:
      description: Calendar or event was not found, or is not visible to the authenticated user.
      content:
        text/plain; charset=utf-8:
          schema:
            $ref: "#/components/schemas/ErrorText"
    Conflict:
      description: The request conflicts with existing calendar data or resource constraints.
      content:
        text/plain; charset=utf-8:
          schema:
            $ref: "#/components/schemas/ErrorText"
    PreconditionFailed:
      description: Conditional request headers were not satisfied.
      content:
        text/plain; charset=utf-8:
          schema:
            $ref: "#/components/schemas/ErrorText"
...
    InternalServerError:
      description: Unexpected server error.
      content:
        text/plain; charset=utf-8:
          schema:
            $ref: "#/components/schemas/ErrorText"
  schemas:
...
    ErrorText:
      type: string
      description: Plain-text error response produced by `http.Error`.
    AddressBook:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    Contact:
      type: object
      additionalProperties: false
      required:
        - uid
        - addressBookId
        - resourceName
        - etag
        - lastModified
        - rawVcard
      properties:
        uid:
          type: string
        addressBookId:
          type: integer
          format: int64
        resourceName:
          type: string
          description: Stored resource name for the contact object.
          example: example-contact.vcf
        displayName:
          type: string
        email:
          type: string
          description: Primary email address parsed from the vCard.
        birthday:
          type: string
          format: date
        etag:
          type: string
        lastModified:
          type: string
          format: date-time
        rawVcard:
          type: string
          description: Raw vCard data for the contact.
    ContactWriteRequest:
      type: object
      additionalProperties: false
      properties:
        inputMode:
          type: string
          enum:
            - structured
            - raw_vcard
          description: Defaults to `structured` when omitted.
        rawVcard:
          type: string
          description: Required when `inputMode` is `raw_vcard`.
        structured:
          $ref: "#/components/schemas/StructuredContactInput"
    StructuredContactInput:
      type: object
      additionalProperties: false
      description: Structured contact fields assembled into a vCard server-side.
      required:
        - displayName
      properties:
        uid:
          type: string
          description: Optional. Generated when omitted; on update must match the path UID.
        displayName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        birthday:
          type: string
          description: ISO date (YYYY-MM-DD) or vCard --MM-DD form.
        notes:
          type: string
        company:
          type: string
    RawVCard:
      type: string
      description: Raw vCard (VCARD) data.
      example: |
        BEGIN:VCARD
        VERSION:3.0
        UID:example-contact
        FN:Example Person
        EMAIL;TYPE=INTERNET:person@example.com
        END:VCARD